--- components/Backdrop/Backdrop.module.scss
@@ -1,9 +1,0 @@
-@import "../../globalStyles/variables.scss";
-.backdrop {
-	width: 100%;
-	height: 100vh;
-	position: fixed;
-	top: 0;
-	left: 0;
-	background-color: rgba(40, 124, 128, 0.77);
-}
