--- components/Modals/Modal.module.scss
@@ -1,93 +1,0 @@
-@import "../../globalStyles/variables.scss";
-.modal {
-	position: fixed;
-	margin: 0 auto;
-	padding: 32px 25px;
-	top: 18vh;
-	left: calc((100% - 600px) / 2);
-	width: 600px;
-	border-radius: 8px;
-	box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.78);
-	border: solid 1px $midGrey;
-	background-color: $white;
-
-	h1 {
-		font-family: $subheadFont;
-		font-size: 16px;
-		font-weight: 600;
-		font-stretch: normal;
-		font-style: normal;
-		line-height: normal;
-		letter-spacing: normal;
-		color: $darkGrey;
-	}
-	form {
-		display: grid;
-		grid-template-rows: 1fr 1fr 1fr;
-		grid-template-columns: 1fr 1fr;
-		height: 100%;
-
-		input {
-			width: 100%;
-			height: 36px;
-			border-radius: 8px;
-			border: solid 1px $midGrey;
-			padding: 3px 10px;
-			font-family: $subheadFont;
-			color: $black;
-		}
-		label {
-			padding-left: 8px;
-			font-family: $headFont;
-			font-size: 13px;
-			font-weight: normal;
-			font-stretch: normal;
-			font-style: normal;
-			line-height: normal;
-			letter-spacing: normal;
-			color: $black;
-		}
-		.email {
-			grid-row: 1 / 2;
-			grid-column: 1 / 3;
-		}
-		.password {
-			grid-row: 2 / 3;
-			grid-column: 1 / 3;
-		}
-		.switch {
-			grid-row: 3 / 4;
-			grid-column: 1 / 2;
-			align-self: center;
-			font-family: $headFont;
-			font-size: 13px;
-			font-weight: normal;
-			font-stretch: normal;
-			font-style: normal;
-			line-height: normal;
-			letter-spacing: normal;
-			color: $darkGrey;
-			.bold {
-				font-weight: 900;
-				cursor: pointer;
-			}
-		}
-		.submit {
-			grid-row: 3 / 4;
-			grid-column: 2 / 3;
-			align-self: center;
-			justify-self: end;
-			width: 106px;
-			height: 36px;
-			border: none;
-			border-radius: 5px;
-			background-color: $brightOrange;
-			font-family: $subheadFont;
-			font-size: 16px;
-			font-weight: 600;
-			text-align: center;
-			color: $white;
-			cursor: pointer;
-		}
-	}
-}
