--- components/Modals/SignUpModal.module.scss
@@ -1,147 +1,0 @@
-@import "../../globalStyles/variables.scss";
-
-.modal {
-	position: absolute;
-	margin: 0 auto;
-	padding: 32px 25px;
-	top: 8vh;
-	left: calc((100% - 600px) / 2);
-	width: 600px;
-	border-radius: 8px;
-	box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.78);
-	border: solid 1px $midGrey;
-	background-color: $white;
-	height: 650px;
-	h1 {
-		font-family: $subheadFont;
-		font-size: 16px;
-		font-weight: 600;
-		font-stretch: normal;
-		font-style: normal;
-		line-height: normal;
-		letter-spacing: normal;
-		color: $darkGrey;
-	}
-	form {
-		display: grid;
-		grid-template-rows: 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr;
-		grid-template-columns: 1fr 1fr;
-		height: 100%;
-
-		input {
-			width: 100%;
-			height: 36px;
-			border-radius: 8px;
-			border: solid 1px $midGrey;
-			padding: 3px 10px;
-			font-family: $subheadFont;
-			color: $black;
-		}
-		input[type="radio"] {
-			opacity: 100;
-			position: absolute;
-		}
-		span {
-			padding-left: 10px;
-			font-size: 10px;
-			color: red;
-		}
-
-		label {
-			padding-left: 8px;
-			font-family: $headFont;
-			font-size: 13px;
-			font-weight: normal;
-			font-stretch: normal;
-			font-style: normal;
-			line-height: normal;
-			letter-spacing: normal;
-			color: $black;
-		}
-		.fName {
-			grid-row: 1 / 2;
-			grid-column: 1 / 3;
-		}
-		.lName {
-			grid-row: 2 / 3;
-			grid-column: 1 / 3;
-		}
-		.email {
-			grid-row: 3 / 4;
-			grid-column: 1 / 3;
-		}
-		.zip {
-			grid-row: 4 / 5;
-			grid-column: 1 / 3;
-		}
-		.password {
-			grid-row: 5 / 6;
-			grid-column: 1 / 3;
-		}
-		.confirmPassword {
-			grid-row: 6 / 7;
-			grid-column: 1 / 3;
-		}
-		.profileType {
-			grid-row: 7 / 8;
-			grid-column: 1 / 3;
-			.buttonContainer {
-				display: flex;
-
-				button {
-					font-family: $headFont;
-					font-size: 13px;
-					height: 24px;
-					padding: 3px 8px 5px;
-					border: none;
-					color: #d9d8d8;
-					background-color: $brownGrey;
-					border-radius: 5px;
-					margin-right: 16px;
-					cursor: pointer;
-					&:focus {
-						outline: none;
-					}
-					&::after {
-						background-color: $brightOrange;
-						color: $white;
-					}
-				}
-			}
-		}
-		.switch {
-			grid-row: 8 / 9;
-			grid-column: 1 / 2;
-			align-self: center;
-			font-family: $headFont;
-			font-size: 13px;
-			font-weight: normal;
-			font-stretch: normal;
-			font-style: normal;
-			line-height: normal;
-			letter-spacing: normal;
-			color: $darkGrey;
-			.bold {
-				font-weight: 900;
-				cursor: pointer;
-			}
-		}
-		.submit {
-			grid-row: 8 / 9;
-			grid-column: 2 / 3;
-			align-self: center;
-			justify-self: end;
-			width: 106px;
-			height: 36px;
-			border: none;
-			border-radius: 5px;
-			background-color: $brightOrange;
-			font-family: $subheadFont;
-			font-size: 16px;
-			font-weight: 600;
-			text-align: center;
-			color: $white;
-			cursor: pointer;
-		}
-	}
-}
