--- pageStyles/resources.module.scss
@@ -1,259 +1,0 @@
-@import "../globalStyles/variables.scss";
-
-.controllers {
-	width: 80%;
-	margin: 40px auto;
-	.unClicked {
-		background-color: $brownGrey;
-	}
-	.clicked {
-		background-color: $brightOrange;
-	}
-
-	h1 {
-		font-family: $subheadFont;
-		font-size: 32px;
-		font-weight: bold;
-		font-style: italic;
-		color: $black;
-		padding-bottom: 16px;
-	}
-	.subtitle {
-		font-family: $headFont;
-		font-size: 16px;
-		color: $black;
-		padding-bottom: 30px;
-	}
-
-	.financialGroup {
-		display: flex;
-		flex-direction: column;
-		align-items: flex-start;
-		justify-content: space-between;
-		height: 130px;
-
-		h2 {
-			font-family: $subheadFont;
-			font-size: 16px;
-			font-weight: bold;
-			color: $black;
-		}
-		// button {
-		// 	cursor: pointer;
-		// 	padding: 3px 8px 5px;
-		// 	border-radius: 5px;
-		// 	background-color: $brownGrey;
-		// 	border: none;
-		// 	font-family: $headFont;
-		// 	font-size: 13px;
-		// 	font-weight: bold;
-		// 	text-align: center;
-		// 	color: $white;
-		// }
-	}
-
-	.divider {
-		width: 110%;
-		height: 1px;
-		margin: 25px 0 25px -5%;
-		background-color: $midGrey;
-	}
-	.rolesGroup {
-		display: flex;
-		flex-direction: row;
-		align-items: center;
-		padding-bottom: 50px;
-		.textClicked {
-			color: $brightOrange;
-		}
-		.textUnClicked {
-			color: $black;
-		}
-		p {
-			font-family: $subheadFont;
-			font-size: 13px;
-			font-weight: 600;
-			cursor: pointer;
-		}
-		.vertDivider {
-			height: 17px;
-			width: 2px;
-			background-color: $black;
-			margin: 0 15px;
-		}
-	}
-	.yearsDiv {
-		width: 530px;
-		font-family: $headFont;
-		font-size: 16px;
-		height: 70px;
-
-		p {
-			margin-left: 289px;
-			color: $brightOrange;
-		}
-		label {
-			display: flex;
-			width: 530px;
-			flex-direction: row;
-			justify-content: space-between;
-			align-items: center;
-
-			input[type="range"] {
-				width: 375px;
-				margin: 13px 0;
-				background-color: transparent;
-				-webkit-appearance: none;
-			}
-			input[type="range"]:focus {
-				outline: none;
-			}
-			input[type="range"]::-webkit-slider-runnable-track {
-				background: rgba(120, 120, 120, 0.78);
-				border: 0;
-				border-radius: 0.8px;
-				width: 100%;
-				height: 2px;
-				cursor: pointer;
-			}
-			input[type="range"]::-webkit-slider-thumb {
-				margin-top: -13px;
-				width: 28px;
-				height: 28px;
-				background: #ffffff;
-				border: 0.2px solid rgba(0, 0, 0, 0.7);
-				border-radius: 50px;
-				cursor: pointer;
-				-webkit-appearance: none;
-			}
-			input[type="range"]:focus::-webkit-slider-runnable-track {
-				background: #d9d9d9;
-			}
-			input[type="range"]::-moz-range-track {
-				background: rgba(120, 120, 120, 0.78);
-				border: 0;
-				border-radius: 0.8px;
-				width: 100%;
-				height: 2px;
-				cursor: pointer;
-			}
-			input[type="range"]::-moz-range-thumb {
-				width: 28px;
-				height: 28px;
-				background: #ffffff;
-				border: 0.2px solid rgba(0, 0, 0, 0.7);
-				border-radius: 50px;
-				cursor: pointer;
-			}
-			input[type="range"]::-ms-track {
-				background: transparent;
-				border-color: transparent;
-				border-width: 14.4px 0;
-				color: transparent;
-				width: 100%;
-				height: 2px;
-				cursor: pointer;
-			}
-			input[type="range"]::-ms-fill-lower {
-				background: #171717;
-				border: 0;
-				border-radius: 1.6px;
-			}
-			input[type="range"]::-ms-fill-upper {
-				background: rgba(120, 120, 120, 0.78);
-				border: 0;
-				border-radius: 1.6px;
-			}
-			input[type="range"]::-ms-thumb {
-				width: 28px;
-				height: 28px;
-				background: #ffffff;
-				border: 0.2px solid rgba(0, 0, 0, 0.7);
-				border-radius: 50px;
-				cursor: pointer;
-				margin-top: 0px;
-				/*Needed to keep the Edge thumb centred*/
-			}
-			input[type="range"]:focus::-ms-fill-lower {
-				background: rgba(120, 120, 120, 0.78);
-			}
-			input[type="range"]:focus::-ms-fill-upper {
-				background: #d9d9d9;
-			}
-			/*TODO: Use one of the selectors from https://stackoverflow.com/a/20541859/7077589 and figure out
-				  how to remove the virtical space around the range input in IE*/
-			@supports (-ms-ime-align: auto) {
-				/* Pre-Chromium Edge only styles, selector taken from hhttps://stackoverflow.com/a/32202953/7077589 */
-				input[type="range"] {
-					margin: 0;
-					/*Edge starts the margin from the thumb, not the track as other browsers do*/
-				}
-			}
-
-			span {
-				font-size: 13px;
-				font-weight: normal;
-				font-stretch: normal;
-				font-style: normal;
-				line-height: normal;
-				letter-spacing: normal;
-				color: $brownGrey;
-			}
-		}
-	}
-
-	.filtersAndSortContainer {
-		display: flex;
-		flex-direction: row;
-		align-items: center;
-		justify-content: space-between;
-
-		.filterContainer {
-			display: flex;
-			flex-direction: row;
-			align-items: center;
-			justify-content: space-between;
-			width: 530px;
-
-			p {
-				font-size: 16px;
-				font-family: $headFont;
-				color: $black;
-			}
-		}
-
-		.sortContainer {
-			display: flex;
-			flex-direction: row;
-			width: 25%;
-			align-items: center;
-			justify-content: flex-end;
-			font-family: $headFont;
-			font-size: 14px;
-			.textClicked {
-				color: $brightOrange;
-			}
-			.textUnClicked {
-				color: $black;
-			}
-			p {
-				cursor: pointer;
-			}
-
-			span {
-				color: $brownGrey;
-				margin-right: 20px;
-			}
-			.vertDividerThin {
-				height: 15px;
-				width: 1px;
-				background-color: $black;
-				margin: 0 15px;
-			}
-		}
-	}
-}
-.resultsContainer {
-	width: 80%;
-	margin: 40px auto;
-}
