--- pages/api/test.js
@@ -1,7 +1,0 @@
-import dbConnect from '../../utils/dbConnect';
-
-dbConnect();
-
-export default async (req, res) => {
-  res.json({ test: 'Db connected' });
-};
