--- pages/connect.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import Layout from '../components/Layout/Layout.component';
-
-const Connect = () => {
-  return (
-    <Layout>
-      <div>
-        <h1>CONNECT</h1>
-      </div>
-    </Layout>
-  );
-};
-
-export default Connect;
