+++ src/common/components/Footer.module.scss
@@ -1,0 +1,82 @@
+@import "../styles/variables.scss";
+
+.footerContainer {
+	width: 100%;
+	height: 255px;
+	display: grid;
+	grid-template-rows: 1fr;
+	grid-template-columns: 1fr 1fr 1fr;
+	padding: 30px 50px;
+	background-color: $black;
+	color: $white;
+	h2 {
+		font-family: $headFont;
+		font-size: 13px;
+		font-weight: bold;
+		font-stretch: normal;
+		font-style: normal;
+		line-height: normal;
+		letter-spacing: normal;
+	}
+	p {
+		font-family: $subheadFont;
+		font-size: 13px;
+	}
+}
+
+.logoContainer {
+	align-items: left;
+	padding-left: 50px;
+	img {
+		width: 150px;
+		padding-bottom: 13px;
+	}
+	p {
+		font-family: $subheadFont;
+		font-size: 16px;
+		font-weight: bold;
+	}
+}
+.contactContainer {
+	text-align: center;
+	h2 {
+		padding-bottom: 12px;
+	}
+}
+.subscribeContainer {
+	form {
+		display: grid;
+		grid-template-columns: 1fr;
+		grid-template-rows: 1fr 1fr 1fr 1fr;
+		grid-gap: 8px;
+		width: 232px;
+		float: right;
+
+		input {
+			height: 36px;
+			border-radius: 5px;
+			border: solid 0.5px $white;
+			background-color: $black;
+			color: $white;
+			padding: 10px 30px;
+			font-size: 13px;
+			font-family: $subheadFont;
+		}
+		button {
+			cursor: pointer;
+			height: 36px;
+			border-radius: 5px;
+			border: solid 2px $white;
+			font-family: $subheadFont;
+			font-size: 16px;
+			font-weight: 600;
+			font-stretch: normal;
+			font-style: normal;
+			line-height: normal;
+			letter-spacing: normal;
+			text-align: center;
+			color: $white;
+			background-color: $black;
+		}
+	}
+}
