+++ src/common/utils/passwordValidator.js
