+++ src/pages/api/test.js
@@ -1,0 +1,7 @@
+import dbConnect from '../../server/database/dbConnect';
+
+dbConnect();
+
+export default async (req, res) => {
+  res.json({ test: 'Db connected' });
+};
