+++ src/pages/app.scss
