+++ src/pages/connect.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import Layout from '../common/components/Layout';
+
+const Connect = () => {
+  return (
+    <Layout>
+      <div>
+        <h1>CONNECT</h1>
+      </div>
+    </Layout>
+  );
+};
+
+export default Connect;
