+++ src/pages/index.module.scss
@@ -1,0 +1,83 @@
+@import "../common/styles/variables.scss";
+
+.heroContainer {
+	background-image: url("../../public/assets/landing-hero.jpg");
+	background-size: cover;
+	background-repeat: no-repeat;
+	background-position: center;
+	height: 320px;
+	width: 100%;
+
+	.overlay {
+		background-color: rgba(74, 74, 74, 0.3);
+		width: 100%;
+		height: 100%;
+	}
+	h1 {
+		width: 100%;
+		margin: 0 auto;
+		padding: 90px 88px 0;
+		font-family: $subheadFont;
+		font-size: 48px;
+		font-weight: 600;
+		font-style: italic;
+		line-height: normal;
+		letter-spacing: normal;
+		color: $white;
+		z-index: 20;
+	}
+}
+.covidSupport {
+	width: 100%;
+	margin: 0 auto;
+	padding: 50px 90px;
+	background-color: $uglyBlue;
+
+	h2 {
+		font-family: $subheadFont;
+		font-size: 16px;
+		font-weight: bold;
+		font-stretch: normal;
+		font-style: normal;
+		line-height: normal;
+		letter-spacing: normal;
+		color: $white;
+		padding-bottom: 24px;
+	}
+	p,
+	ul,
+	li {
+		font-family: $headFont;
+		font-size: 16px;
+		font-weight: normal;
+		font-stretch: normal;
+		font-style: normal;
+		line-height: normal;
+		letter-spacing: normal;
+		color: $white;
+	}
+	p {
+		padding-bottom: 54px;
+	}
+	ul {
+		p {
+			padding-bottom: 0;
+		}
+		li {
+			list-style: none;
+		}
+	}
+	button {
+		width: 141px;
+		height: 40px;
+		color: $white;
+		background-color: $brightOrange;
+		border-radius: 5px;
+		border: none;
+		font-size: 16px;
+		font-family: $subheadFont;
+		font-weight: 600;
+		margin: 32px 0 57px;
+		cursor: pointer;
+	}
+}
