+++ src/resources/components/ResourceCard.jsx
