+++ src/resources/utils/resources-for-offline.js
