+++ src/server/models/resource.js
